Split project
Split the project between a core and a bot library.
Core only contains the minimum to interface with the different sites.
Bot contains the logic for botting.
the bot library will be moved into it's own github repo as some point, but to make it easier for now, i'll keep it in one repo for now.
DoormatController will also be moved into a seperate github repo with the rest of the UI components